Move standard answers web UI to ee
