+++ web/src/app/ee/admin/standard-answer/new/page.tsx
@@ -1,5 +1,5 @@
 import { AdminPageTitle } from "@/components/admin/Title";
-import { StandardAnswerCreationForm } from "@/app/admin/standard-answer/StandardAnswerCreationForm";
+import { StandardAnswerCreationForm } from "@/app/ee/admin/standard-answer/StandardAnswerCreationForm";
 import { fetchSS } from "@/lib/utilsSS";
 import { ErrorCallout } from "@/components/ErrorCallout";
 import { BackButton } from "@/components/BackButton";
