+++ web/src/middleware.ts
@@ -9,6 +9,7 @@ const eePaths = [
   "/admin/performance/query-history",
   "/admin/whitelabeling",
   "/admin/performance/custom-analytics",
+  "/admin/standard-answer",
 ];
 
 const eePathsForMatcher = eePaths.map((path) => `${path}/:path*`);
